refactor(job): module cleanup + docs for the commit-hook decomposition
Moves ClaimedRow/claim_due_heads_in_op from spawner.rs into poller.rs:
they are claim machinery (used only by ClaimHook and ClaimReconciler),
not insert machinery, and their presence in spawner.rs was the last
piece of the spawner<->poller import cycle the addendum set out to
break. poller.rs no longer imports anything from spawner.rs; the
resulting import graph is spawner/keyed/dispatcher/batch_dispatcher/
resident -> {execution_hooks, poller}, execution_hooks <-> poller
(the one narrow, accepted two-way edge -- ClaimHook needs
PromoteHeadsHook's TypeId for runs_after and calls PromoteHeadsHook::
apply from reclaim_lost_jobs; ExecutionInsertHook needs PollerHandle).

PERFORMANCE.md: added a "Commit-hook pipeline" section naming the
three hooks (ExecutionInsertHook, PromoteHeadsHook, ClaimHook) and
how their ordering is enforced (runs_after plus, within
ExecutionInsertHook's own pre_commit, plain sequential Rust code).
Inverted the old "cost worth naming" lock-hold warning: a claimed
head's row lock is now held only from the claim to COMMIT, not for
the remainder of whatever transaction the caller's write runs in,
since the claim itself now runs inside the commit pass rather than
inline mid-transaction. Documented the real cost that moved the other
way: an insert's errors now surface at op.commit() rather than at the
`_in_op` call site. Restated, more strongly, why Promote+Claim (or
Insert+Promote) must never collapse into one query -- CTEs in a single
statement share one snapshot, so a row promoted in one CTE is invisible
to a sibling CTE in the SAME statement; this is not hypothetical, it is
the exact shape of two real bugs earlier in this branch's history. Old
per-call-site cost note kept in a <details> block for history.

No `@@` review markers remain anywhere in the tree.

Full 142-test suite passes; nix flake check clean.
